+++ shared/src/config.rs
@@ -11,13 +11,11 @@ fn read_file(file_name: &str) -> io::Result<String> {
 
 pub fn parse_server_config(file_name: &str) -> ServerConfig {
     let content = read_file(file_name).expect("Error reading config file");
-    let config = toml::from_str(&content).expect("Error parsing config file");
-    return config
+    toml::from_str(&content).expect("Error parsing config file")
 }
 
 // TODO: This is wet code
 pub fn parse_client_config(file_name: &str) -> ClientConfig {
     let content = read_file(file_name).expect("Error reading config file");
-    let config = toml::from_str(&content).expect("Error parsing config file");
-    return config
+    toml::from_str(&content).expect("Error parsing config file")
 }
