+++ srps/src/main.rs
@@ -1,13 +1,12 @@
-use shared;
 use log::{info};
-use quinn::{Endpoint};
-use quinn::ServerConfig;
 use std::{
     error::Error,
     net::{IpAddr, SocketAddr},
-    sync::Arc,
 };
-use rustls::pki_types::{CertificateDer, PrivatePkcs8KeyDer};
+
+mod configuration;
+
+use configuration::make_server_endpoint;
 
 async fn run_server(addr: SocketAddr) {
     let (endpoint, _server_cert) = make_server_endpoint(addr).unwrap();
@@ -18,28 +17,6 @@ async fn run_server(addr: SocketAddr) {
         "[server] connection accepted: addr={}",
         conn.remote_address()
     );
-}
-
-pub(crate) fn make_server_endpoint(
-    bind_addr: SocketAddr,
-) -> Result<(Endpoint, CertificateDer<'static>), Box<dyn Error + Send + Sync + 'static>> {
-    let (server_config, server_cert) = configure_server()?;
-    let endpoint = Endpoint::server(server_config, bind_addr)?;
-    Ok((endpoint, server_cert))
-}
-
-fn configure_server()
--> Result<(ServerConfig, CertificateDer<'static>), Box<dyn Error + Send + Sync + 'static>> {
-    let cert = rcgen::generate_simple_self_signed(vec!["localhost".into()]).unwrap();
-    let cert_der = CertificateDer::from(cert.cert);
-    let priv_key = PrivatePkcs8KeyDer::from(cert.signing_key.serialize_der());
-
-    let mut server_config =
-        ServerConfig::with_single_cert(vec![cert_der.clone()], priv_key.into())?;
-    let transport_config = Arc::get_mut(&mut server_config.transport).unwrap();
-    transport_config.max_concurrent_uni_streams(0_u8.into());
-
-    Ok((server_config, cert_der))
 }
 
 #[tokio::main]
